% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayz.R
\name{augment.loo}
\alias{augment.loo}
\alias{glance.loo}
\alias{tidy.loo}
\title{Tidying methods for}
\usage{
augment.loo(x, data = NULL, ...)

glance.loo(x, ...)

tidy.loo(x, ...)
}
\arguments{
\item{x}{an \code{\link[loo]{loo}} object}

\item{data}{Data frame with the original data}

\item{...}{Not used currently}
}
\value{
\code{augment.loo} returns a data frame, with one row per observation and the pointwise elpd, elpd standard error, and estimated number of observations.

\code{glance.loo} returns a one row data frame with the columns
  \item{\code{elpd_loo},\code{elpd_waic}}{Expected log pointwise predictive density}
  \item{\code{se_elpd_loo},\code{se_elpd_waic}}{Standard error of the log pointwise predictive density}
  \item{\code{se_looic}, \code{see_waic}}{LOO information criterion (\code{-2 * elpd_loo}), or WAI, i.e. converted to the deviance scance}
  \item{\code{p_loo}, \code{p_waic}}{Estimated effective number of parameters}
  \item{\code{se_p_loo}, \code{se_p_waic}}{Standard error of the estimated effective number of parameters}
  \item{\code{n}}{Number of observations}
  \item{\code{n_sims}}{Number of MCMC sims}

A data frame with columns:
   \item{\code{parameter}}{Paramter: elpd_loo or elpd_waic, p_loo or p_waic, or waic or looic}
   \item{\code{estimate}}{Parameter value}
   \item{\code{std.error}}{Standard error of the parameter}
}
\description{
Methods for tidying the results from from WAIC and LOO cross-validation methods in the \pkg{loo} package.
}
