% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscale.R
\name{autoscale}
\alias{autoscale}
\alias{autoscale.numeric}
\alias{autoscale.logical}
\alias{autoscale.matrix}
\alias{autoscale.data.frame}
\title{Center and Scale}
\usage{
autoscale(x, ...)

\method{autoscale}{numeric}(x, center = TRUE, scale = TRUE, ...)

\method{autoscale}{logical}(x, center = TRUE, scale = TRUE, ...)

\method{autoscale}{matrix}(x, center = TRUE, scale = TRUE, ...)

\method{autoscale}{data.frame}(x, center = TRUE, scale = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{...}{Other arguments (not used)}

\item{center}{If \code{TRUE}, then set the mean of the predictor
to zero.}

\item{scale}{If \code{TRUE}, then scale the predictor.}
}
\description{
Center and scale predictors using the same methods
as in \pkg{rstanarm}.
}
\details{
Variables are standardized using the following rules, based on the
number of unique values:
\itemize{
\item{one: nothing}
\item{two: subtract the mean and divide by the range}
\item{three+: subtract the mean and divide by the standard deviation}
}
}
