% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{dhs}
\alias{dhs}
\alias{rhs}
\title{The Horseshoe Prior Distribution}
\usage{
dhs(x, tau = 1, location = 0)

rhs(n, tau = 1, location = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles}

\item{tau}{numeric vector. The scale parameter}

\item{location}{numeric vector. The location parameter.}

\item{n}{Number of samples to draw}
}
\value{
A numeric vector
}
\description{
The Horseshoe Prior Distribution
}
\references{
Carvalho, C. M., Polson, N. G., Scott, J. G. (2010)
   \dQuote{The horseshoe estimator for sparse signals}. \emph{Biometrika}
   \href{https://dx.doi.org/10.1093/biomet/asq017}{doi:10.1093/biomet/asq017}.
}
