% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsplus.R
\name{dhsplus}
\alias{dhsplus}
\alias{rhsplus}
\title{The Horseshoe+ Prior Distribution}
\usage{
dhsplus(x, tau = 1, location = 0)

rhsplus(n, tau = 1, location = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles}

\item{tau}{numeric vector. The scale parameter}

\item{location}{numeric vector. The location parameter.}

\item{n}{Number of samples to draw}
}
\value{
A numeric vector
}
\description{
The Horseshoe+ Prior Distribution
}
\details{
There is no exact closed form for the density of the horseshoe
distribution.

The Horseshoe+ Prior Distribution is defined hierarchically as,
\deqn{
x_i | \lambda_i \sim N(0, \lambda_i^2)
\lambda_i | \eta_i \sim C^{+}(0, \eta_i)
\eta_i | \eta_i \sim C^{+}(0, 1)
}
Integrating over \eqn{\eta_i}, the density of \eqn{\lambda_i} is,
\deqn{
p(\lambda_i) = \frac{4}{\pi^2} \frac{\log(\lambda_i)}{\lambda_i^2 - 1} .
}

There is no closed form representation of the density, but it is bounded by,
\deqn{
\frac{1}{\pi^2 \sqrt{2 \pi}} \log \left( 1 + \frac{4}{\theta^2} \right)
< p_{HS+}(\theta)
\leq \frac{1}{\pi^2 | |\theta|}
}{
1 / (pi^2 * sqrt(2 pi)) * log(1 + 4 / theta^2)
< p_HS+ (theta)
<=  1 / (pi^2 * abs(theta))
}
}
\references{
Bhadra, Anindya; Datta, Jyotishka; Polson, Nicholas G.; Willard, Brandon. The Horseshoe+ Estimator of Ultra-Sparse Signals. Bayesian Anal., advance publication, 22 September 2016. doi: 10.1214/16-BA1028. http://projecteuclid.org/euclid.ba/1474572263
}
