% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeffreys.R
\name{djeffreys}
\alias{djeffreys}
\alias{rjeffreys}
\title{Jeffreys' Variance Reference Prior}
\usage{
djeffreys(x)

rjeffreys(n)
}
\arguments{
\item{x}{A numeric vector of quantiles}

\item{n}{The number of samples to draw}
}
\description{
The Jeffreys' improper prior for variance is
\eqn{p(x) \propto 1 / x}. This is not a proper probability distribution.
However, "density" and random sampling functions are provided.
}
\details{
The Jeffreys' improper prior can be approximated by a Gamma distribution
as the shape and rate go to zero.
}
