% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_functions.R
\name{kappa_lambda_funs}
\alias{kappa_lambda_funs}
\alias{kappa_tau_funs}
\alias{kappa_sigma_funs}
\title{Shinkage parameter functions}
\usage{
kappa_lambda_funs(tau = 1, sigma = 1, n = 1)

kappa_tau_funs(lambda = 1, sigma = 1, n = 1)

kappa_sigma_funs(lambda = 1, tau = 1, n = 1)
}
\arguments{
\item{tau}{numeric. The global scale parameter.}

\item{sigma}{numeric. The observation disturbance scale parameter.}

\item{n}{integer. The number of observations.}

\item{lambda}{numeric. The local scale parameter.}
}
\value{
A list of functions
}
\description{
Generate functions of the shrinkage parameter,
\deqn{
\kappa_j = \frac{1}{1 + n \sigma^{-1} \tau^2 \lambda_j^2} ,
}
with respect to lambda, sigma, or tau, as well as the inverse functions
and derivatives of the inverse functions (with respect to scale, variance, or precisition).
These functions can be used in a transformation of variables, to generate
the density of the probability of distribution of kappa given the probability
distributuion of lambda, sigma, or tau.
}
