% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_moments.R
\name{kappa_moments}
\alias{kappa_moments}
\alias{meff_moments}
\title{Moments for shrinkage and number of effective parameters}
\usage{
kappa_moments(tau, sigma = 1, n = 1)

meff_moments(tau, sigma = 1, n = 1, D = 1)
}
\arguments{
\item{tau}{Numeric. The global scale parameter.}

\item{sigma}{Numeric. The observation disturbance scale parameter.}

\item{n}{Integer. The number of observations.}

\item{D}{Integer. The number of parameters.}
}
\value{
\describe{
\item{kappa_moments}{A data frame with columns: \code{tau}, \code{sigma}, \code{n},
      \code{mean}, and \code{var}.}
\item{meff_moments}{A data frame with columns: \code{tau}, \code{sigma}, \code{n},
      \code{D}, \code{mean}, and \code{var}.}
}
}
\description{
Means and variances for the shrinkage parameter, \eqn{kappa}, and
the number of effective parameters, \eqn{m_{eff}}, in regression
problems using scale-mixture of normal shrinkage priors.
}
\details{
The mean and variance for the shrinkage parameter \eqn{\kappa} given the global scale
\eqn{\tau}, and observation disturbance scale \eqn{\sigma} are:
\eqn{
\begin{aligned}[t]\
E(\kappa_j | \tau, \sigma) &= \frac{1}{1 + \sigma^{-1} \tau \sqrt{n}} , \\
Var(\kappa_j | \tau, \sigma) &= \frac{\sigma^{-1} \tau \sqrt{n}}{2(1 + \sigma^{-1} \tau \sqrt{n})^2} \\
\end{aligned}
}
Likewise, the mean and variance for the number of effective parameters \eqn{m_{eff}}
for \eqn{D} parameters is,
\eqn{
\begin{aligned}[t]
E(m_{eff} | \tau, \sigma) &= \frac{1}{1 + \sigma^{-1} \tau \sqrt{n}} D, \\
Var(m_{eff} | \tau, \sigma) &= \frac{\sigma^{-1} \tau \sqrt{n}}{2(1 + \sigma^{-1} \tau \sqrt{n})^2} D . \\
\end{aligned}
}
}
