% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkjcorr.R
\name{rlkjcorr}
\alias{rlkjcorr}
\title{LKJ Correlation Distribution}
\usage{
rlkjcorr(n, eta = 1, cholesky = FALSE, permute = !cholesky)
}
\arguments{
\item{n}{Size of the matrix}

\item{eta}{The shape parameter}

\item{cholesky}{If \code{TRUE}, return the Cholesky distribution}

\item{permute}{If \code{TRUE}, permute the rows.}
}
\value{
A matrix
}
\description{
LKJ Correlation Distribution
}
