% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayz.R
\name{stan_iter}
\alias{stan_iter}
\title{Coerce \code{stanfit} object to list of iterations}
\usage{
stan_iter(object, permuted = FALSE, inc_warmup = FALSE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{stanfit}}.}

\item{permuted}{A logical scalar indicating whether the draws
    after the \emph{warmup} period in each chain should be \emph{permuted} and 
    \emph{merged}. If \code{FALSE}, the original order is kept. For each 
    \code{stanfit} object, the permutation is fixed (i.e., extracting samples 
    a second time will give the same sequence of iterations).}

\item{inc_warmup}{A logical scalar indicating whether to include
    the warmup draws. This argument is only relevant if \code{permuted} is 
    \code{FALSE}.}
}
\value{
A list. Each element is a named list representing an iteration.
 The names of the list which the names are parameters,
 and the elements are numeric arrays with the parameters in their original dimensions.
}
\description{
Extract the parameters from a \code{\link[rstan]{stanfit}} object and coerce them into
a list in which each element is a list of the parameters in their original dimenstions.
}
