% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-stan-workflow.R
\name{stan_workflow}
\alias{stan_workflow}
\alias{check_div}
\alias{check_treedepth}
\alias{check_energy}
\alias{partition_div}
\title{Robust Stan Workflow Functions}
\usage{
check_div(fit)

check_treedepth(fit)

check_energy(fit, threshhold = 0.2)

partition_div(fit, pars = NULL)
}
\arguments{
\item{fit}{A \code{stanfit} object.}

\item{threshhold}{If any chain has an E-BFI below \code{threshshold}, a
warning is printed.}

\item{pars}{A character vector of parameters to include.}
}
\value{
These functions are largely run interactively in order to
\describe{
  \item{check_div}{A list with two elements: \code{n_div} is an integer scalar
    with the total number of divergences, and \code{n_iter} is an integer scalar
    with the total number of iterations.}
  \item{check_treedepth}{A list with three elements:
   \describe{
   \item{\code{max_depth}}{The maximum treedepth used in the sampler.}
   \item{\code{n_max}}{The number of iterations reaching the maximum depth.}
   \item{\code{n_iter}}{The total number of elements.}
   }}
  \item{check_energy}{A numeric vector with the E-BFI for each chain.}
  \item{partition_div}{A data frame with a row for each (sample, parameter),
    and columns,
    \describe{
    \item{\code{.iter}}{Iteration number}
    \item{\code{.chain}}{Chain number}
    \item{\code{.div}}{Was this iteration a divergence?}
    \item{\code{parameter}}{parameter name}
    \item{\code{value}}{parameter value}
    }}
}
}
\description{
These functions are derived from from Michael Betancourt's \href{https://github.com/betanalpha/knitr_case_studies/blob/master/rstan_workflow/stan_utility.R}{Robust Stan Workflow},
Stan case study.
}
\details{
\describe{
\item{check_div}{Check transitions that ended with a divergence}
\item{check_energy}{Check chains for low E-BFI.}
\item{check_treedepth}{}
\item{partition_div}{Partitions parameters into divergent and non-divergent iterations.}
}
}
