% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit-summary.R
\name{stanfit_summary}
\alias{stanfit_summary}
\alias{print.stanfit_summary}
\alias{as_data_frame.stanfit_summary}
\title{Class for \code{stanfit} summary objects}
\usage{
stanfit_summary(x)

\method{print}{stanfit_summary}(x, ..., n = NULL, width = NULL,
  n_extra = NULL, stats = c("n_eff", "Rhat", "mean", "se_mean", "sd"))

\method{as_data_frame}{stanfit_summary}(x, summary = TRUE, chains = FALSE,
  ...)
}
\arguments{
\item{x}{For \code{stanfit_summary}, the result of the \code{stanfit}
\code{summary} method. For methods, a \code{stanfit_summary} object.}

\item{...}{Not used.}

\item{n}{Number of rows to print}

\item{width}{Width of text output to generate.}

\item{n_extra}{Number of extra columns for which to print abbreviated
information.}

\item{stats}{character vector with the names of the stats to include.}

\item{summary}{If \code{TRUE}, then return the summary of all chains.}

\item{chains}{If \code{TRUE}, then return the individual chains. One of
\code{summary} or \code{chains} must be \code{TRUE}. They will be appended
together.}
}
\value{
An object of class \code{stanfit_summary}
}
\description{
Class for \code{stanfit} summary objects
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print the summary object

\item \code{as_data_frame}: Converts the summary object to a data frame.
}}

