% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_scale.R
\name{tau_prior_scale}
\alias{tau_prior_scale}
\title{Global scale prior scale choice}
\usage{
tau_prior_scale(p, D = 1, n = D, sigma = 1)
}
\arguments{
\item{p}{Expected number of non-zero paramters}

\item{D}{Number of parameters}

\item{n}{Number of observations}

\item{sigma}{Scale of the observation disturbances}
}
\value{
A numeric vector
}
\description{
Global scale prior scale choice
}
\details{
Piironen and Vehtari (2017) suggest the following choice of the scale for the
global hyperparameter in a regression shrinkage problem with a global-local
scale mixture of normal prior distribution,
\deqn{
\tau_0 = \frac{p}{D - p} \frac{\sigma}{\sqrt{n}}
}
where \eqn{p} is the expected number of non-zero paramters, \code{sigma}
is the scale of the observation disturbances, \code{D} is the number of
parameters, and \code{n} is the number of observations.
}
\references{
Piironen, Juho, and Aki Vehtari. 2017. "On the Hyperprior Choice for the Global Shrinkage Parameter
in the Horseshoe Prior" aXiv:1610.05559v2
}
