% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unflatten_pars.R
\name{unflatten_pars}
\alias{unflatten_pars}
\title{Unflatten Stan parameters}
\usage{
unflatten_pars(object)
}
\arguments{
\item{object}{A \code{"\link[=stanfit-class]{stanfit}"} object.}
}
\value{
A function that takes a single argument \code{x} which
 is a expected to be a numeric vector.
}
\description{
This function returns a function that will convert of vector of flattened
parameters from the output of a Stan model into a list with the original
parameters with their original values.
}
