% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{dhs}
\alias{dhs}
\alias{rhs}
\title{The Horseshoe Prior Distribution}
\usage{
dhs(x, tau = 1, location = 0)

rhs(n, tau = 1, location = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles}

\item{tau}{numeric vector. The scale parameter}

\item{location}{numeric vector. The location parameter.}

\item{n}{Number of samples to draw}
}
\value{
A numeric vector
}
\description{
The Horseshoe Prior Distribution
}
\details{
The Horseshoe Prior Distribution is defined hierarchically as,
\deqn{
x_i | \lambda_i \sim N(0, \lambda_i^2)
\lambda_i | \eta_i \sim C^{+}(0, \eta_i)
}
The density of \eqn{\lambda}{lambda} is a half-Cauchy,
\deqn{
p(\lambda_i) = \frac{2}{\pi^2} \frac{\log(\lambda_i)^2}{\lambda_i^2 - 1} .
}

There is no closed form representation of the density, but it is bounded by,
\deqn{
\frac{1}{\pi^2 \sqrt{2 \pi}} \log \left( 1 + \frac{4}{\theta^2} \right)
< p_{HS+}(\theta)
\leq \frac{1}{\pi^2 | |\theta|}
}{
1 / (pi^2 * sqrt(2 pi)) * log(1 + 4 / theta^2)
< p_HS+ (theta)
<=  1 / (pi^2 * abs(theta))
}
}
\references{
Carvalho, C. M., Polson, N. G., Scott, J. G. (2010)
   \dQuote{The horseshoe estimator for sparse signals}. \emph{Biometrika}
   \href{https://dx.doi.org/10.1093/biomet/asq017}{doi:10.1093/biomet/asq017}.
}
